task-service/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]